+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to ds-license will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+### Changed
+- Upgraded from HttpSolrClient(v8) to Http2SolrClient (v9).
+
+
 ## [4.0.0](https://github.com/kb-dk/ds-license/releases/tag/ds-license-4.0.0) 2026-01-29
 
 ### Added